feat(agent-bus): rename EVA surface to Polly Operator
