+++ packages/agent-bus/src/station-manifest.ts
@@ -747,7 +747,7 @@ export function sweepKeepers(
   };
 }
 
-// ─── EVA station summary ──────────────────────────────────────────────────────
+// ─── Polly station summary ────────────────────────────────────────────────────
 
 export interface DistrictHealthRecord {
   district: District;
